Install vite, typescript , eslint
